Substitute dbt.cc to have advertised licencing #10478
