+++ TARGETS
@@ -328,7 +328,7 @@ cpp_library_wrapper(name="rocksdb_lib", srcs=[
         "utilities/transactions/lock/range/range_tree/lib/locktree/txnid_set.cc",
         "utilities/transactions/lock/range/range_tree/lib/locktree/wfg.cc",
         "utilities/transactions/lock/range/range_tree/lib/standalone_port.cc",
-        "utilities/transactions/lock/range/range_tree/lib/util/dbt.cc",
+        "utilities/transactions/lock/range/range_tree/lib/util/dbt_subst.cc",
         "utilities/transactions/lock/range/range_tree/lib/util/memarena.cc",
         "utilities/transactions/lock/range/range_tree/range_tree_lock_manager.cc",
         "utilities/transactions/lock/range/range_tree/range_tree_lock_tracker.cc",
